\name{CCregPCA-package}
\alias{CCregPCA-package}
\alias{CCregPCA}
\docType{package}
\title{
\packageTitle{CCregPCA}
}
\description{
\packageDescription{CCregPCA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CCregPCA}
\packageIndices{CCregPCA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CCregPCA}

Maintainer: \packageMaintainer{CCregPCA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
